Reorganized files to make them browser friendly for #3.
