+++ safe-access.js
